+++ src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/requests/GetMe.kt
@@ -1,15 +1,14 @@
 package com.github.insanusmokrassar.TelegramBotAPI.requests
 
 import com.github.insanusmokrassar.TelegramBotAPI.requests.abstracts.SimpleRequest
-import com.github.insanusmokrassar.TelegramBotAPI.types.Bot
-import com.github.insanusmokrassar.TelegramBotAPI.types.User
+import com.github.insanusmokrassar.TelegramBotAPI.types.*
 import kotlinx.serialization.*
 
 @Serializable
-class GetMe : SimpleRequest<Bot> {
+class GetMe : SimpleRequest<ExtendedBot> {
     override fun method(): String = "getMe"
-    override val resultDeserializer: DeserializationStrategy<Bot>
-        get() = Bot.serializer()
+    override val resultDeserializer: DeserializationStrategy<ExtendedBot>
+        get() = ExtendedBot.serializer()
     override val requestSerializer: SerializationStrategy<*>
         get() = serializer()
 }