+++ CHANGELOG.md
@@ -29,8 +29,10 @@
         * `QuizKeyboardButtonPollType`
 * `User` now is sealed class
     * `CommonUser` was added as representation of default `User`
-    * `Bot` was added as representation of bot user
-    * `GetMe` now return `Bot` object
+    * `Bot` was added as representation of bot user (it is sealed class)
+        * `ExtendedBot` with additional info
+        * `CommonBot` with simple info
+    * `GetMe` now return `ExtendedBot` object
     * Now extension `javaLocale` is extension for `CommonUser`
 
 ## 0.22.0